requirements.txt
torch==2.5.1
torchvision==0.20.1
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.4
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
streamlit==1.29.0
streamlit-webrtc==0.47.1
albumentations==1.3.1
imgaug==0.4.0
av==11.0.0
PyYAML==6.0.1
tqdm==4.66.1
h5py==3.10.0
tensorboard==2.15.1
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
pytest==7.4.3
jupyter==1.0.0
ipykernel==6.27.1